% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/direct.label.prevR.r
\name{direct.label.prevR}
\alias{direct.label.prevR}
\title{Direct label on a ggplot object}
\usage{
direct.label.prevR(p, method = NULL, debug = FALSE)
}
\arguments{
\item{p}{The ggplot object.}

\item{method}{Method for direct labeling as described in ?label.positions.}

\item{debug}{Show debug output?}
}
\value{
The ggplot object with direct labels added.
}
\description{
Direct label a ggplot2 grouped plot
}
\note{
This function is based on and similar to \code{\link[directlabels]{direct.label.ggplot}}
except that legend is not hidden.
}
\seealso{
\code{\link[directlabels]{direct.label.ggplot}}\{\pkg{directlabels}\}
}

