% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prevR-package.r
\docType{data}
\name{fdhs}
\alias{fdhs}
\alias{fdhs.boundary}
\alias{fdhs.clusters}
\title{Fictitious data generated by a DHS simulation.}
\description{
Data set generated by a Demographic and Health Survey (DHS) simulation on a fictitious
country with a national prevalence of 10\%, 8000 having been surveyed, distributed in 401 clusters.
This dataset is composed of 3 objects:\itemize{
   \item \code{fdhs.clusters}: data frame (one line per cluster).
   \item \code{fdhs.boundary}: object of class
     \code{\link[sp:SpatialPolygons-class]{SpatialPolygons}} corresponding to the borders of the fictitious country.
   \item \code{fdhs}: object of class \code{\link[=prevR-class]{prevR}}
     returned by \code{\link{as.prevR}} using the two previous objects.
}
}
\examples{
\dontrun{
   str(fdhs)
   str(fdhs.clusters)
   str(fdhs.boundary)
   demo(prevR)
 }
}
\keyword{datasets}

