% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noptim.r
\name{Noptim}
\alias{Noptim}
\title{Suggested optimal value for N}
\usage{
Noptim(object)
}
\arguments{
\item{object}{object of class \linkS4class{prevR}.}
}
\value{
an integer.
}
\description{
Based on previous simulation work, the function suggests an optimal
value for the N parameter based on national prevalence, the total number
of observations and the number of clusters. See Larmarange et al. 2011
for more details.
}
\examples{
Noptim(fdhs)

}
\references{
Larmarange Joseph, Vallo Roselyne, Yaro Seydou, Msellati Philippe and Meda Nicolas (2011)
"Methods for mapping regional trends of HIV prevalence from Demographic and Health Surveys (DHS)",
\emph{Cybergeo: European Journal of Geography}, no 558, \url{https://journals.openedition.org/cybergeo/24606},
DOI: 10.4000/cybergeo.24606.
}
\keyword{stat}
