% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_prevR.r
\name{theme_prevR}
\alias{theme_prevR}
\alias{theme_prevR_light}
\title{prevR themes for ggplot2}
\usage{
theme_prevR(base_size = 12)

theme_prevR_light(base_size = 12)
}
\arguments{
\item{base_size}{base font size}
}
\description{
Two custom themes for ggplot2 graphs, hiding axis.
}
