% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prevR.R
\name{update_prevR}
\alias{update_prevR}
\title{Update a prevR object}
\usage{
update_prevR(object)
}
\arguments{
\item{object}{a \code{prevR} object}
}
\value{
a \code{prevR} object
}
\description{
Update an object of class \code{prevR} created with a previous version of the
package to the last version. In particular, it will convert any boundary
slot defined with the \code{sp} package to \code{sf} class.
}
