% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.outside.SpatialPolygons.prevR.r
\name{NA.outside.SpatialPolygons}
\alias{NA.outside.SpatialPolygons}
\title{Affect NA value to points located outside a polygon.}
\usage{
NA.outside.SpatialPolygons(sp.data, sp.poly)
}
\arguments{
\item{sp.data}{object of class \link[sp:SpatialPixelsDataFrame-class]{sp::SpatialPixelsDataFrame}.}

\item{sp.poly}{object of class \link[sp:SpatialPolygons-class]{sp::SpatialPolygons}.}
}
\value{
Return \code{sp.data} modified.
}
\description{
This function forces points of an object of class
\link[sp:SpatialPixelsDataFrame-class]{sp::SpatialPixelsDataFrame} located outside
the limits defined by an object of class \link[sp:SpatialPolygons-class]{sp::SpatialPolygons}
to \code{NA}.
}
\seealso{
\code{\link[=point.in.SpatialPolygons]{point.in.SpatialPolygons()}}.
}
\keyword{manip}
\keyword{spatial}
