% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.prevR.r
\name{print,prevR-method}
\alias{print,prevR-method}
\alias{print}
\alias{print-methods}
\title{Summary of a prevR object.}
\usage{
\S4method{print}{prevR}(x)
}
\arguments{
\item{x}{object of class \linkS4class{prevR}.}
}
\description{
Method \code{print} for objects of class \linkS4class{prevR}:
shows a summary of the object's characteristics.
}
\note{
Exactly the same as \code{\link[=show]{show()}}.
}
\examples{
print(fdhs)
\dontrun{
dhs <- rings(fdhs, N = c(100, 300, 500))
print(dhs)
}

}
\seealso{
\code{\link[=summary]{summary()}}.
}
