% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.prevR.r
\docType{methods}
\name{print,prevR-method}
\alias{print}
\alias{print,prevR-method}
\alias{print-methods}
\title{Summary of a prevR object.}
\usage{
\S4method{print}{prevR}(x)
}
\arguments{
\item{object}{object of class \code{\link[=prevR-class]{prevR}}.}
}
\description{
Method \code{print} for objects of class \code{\link[=prevR-class]{prevR}}:
shows a summary of the object's characteristics.
}
\note{
Exactly the same as \code{\link{show,prevR-method}}.
}
\examples{
print(fdhs)
\dontrun{
  dhs <- rings(fdhs,N=c(100,300,500))
  print(dhs)
}
}
\seealso{
\code{\link{summary,prevR-method}}.
}

