% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.prevR.r
\docType{methods}
\name{show,prevR-method}
\alias{show}
\alias{show,prevR-method}
\alias{show-methods}
\title{Summary of a prevR object.}
\usage{
\S4method{show}{prevR}(object)
}
\arguments{
\item{object}{object of class \code{\link[=prevR-class]{prevR}}.}
}
\description{
Method \code{show} for objects of class \code{\link[=prevR-class]{prevR}}:
shows a summary of the object's characteristics.
}
\note{
Exactly the same as \code{\link{print,prevR-method}}.
}
\examples{
fdhs
\dontrun{
  dhs <- rings(fdhs,N=c(100,300,500))
  dhs
}
}
\seealso{
\code{\link{summary,prevR-method}}.
}

